book-recommendation-system/backend/requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.26
Flask-Cors==5.0.0
